place integration tests in the correct gradle task folder
