+++ airbyte-integrations/connectors/source-postgres/src/test-integration/kotlin/io/airbyte/integrations/source/postgres/PostgresSourceAcceptanceLegacyCtidTest.kt
@@ -2,9 +2,7 @@
  * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.source.postgres.legacy
-
-import io.airbyte.integrations.source.postgres.legacy.testFixtures.PostgresTestDatabase
+package io.airbyte.integrations.source.postgres
 
 class PostgresSourceAcceptanceLegacyCtidTest : PostgresSourceAcceptanceTest() {
     override val serverImage: PostgresTestDatabase.BaseImage
