+++ airbyte-integrations/connectors/source-postgres/src/test-integration/kotlin/io/airbyte/integrations/source/postgres/PostgresSourceDatatypeTest.kt
@@ -2,12 +2,11 @@
  * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.source.postgres.legacy
+package io.airbyte.integrations.source.postgres
 
 import com.fasterxml.jackson.databind.JsonNode
 import io.airbyte.cdk.test.fixtures.legacy.Database
 import io.airbyte.cdk.test.fixtures.legacy.JdbcUtils.MODE_KEY
-import io.airbyte.integrations.source.postgres.legacy.testFixtures.PostgresTestDatabase
 import java.sql.SQLException
 
 class PostgresSourceDatatypeTest : AbstractPostgresSourceDatatypeTest() {
