+++ airbyte-integrations/connectors/source-postgres/src/test-integration/kotlin/io/airbyte/integrations/source/postgres/PostgresSourceSSLCaCertificateAcceptanceTest.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.source.postgres.legacy
+package io.airbyte.integrations.source.postgres
 
 class PostgresSourceSSLCaCertificateAcceptanceTest :
     AbstractPostgresSourceSSLCertificateAcceptanceTest() {
