+++ airbyte-integrations/connectors/source-postgres/src/testFixtures/kotlin/io/airbyte/integrations/source/postgres/PostgresContainerFactory.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2026 Airbyte, Inc., all rights reserved.
  */
 
-package io.airbyte.integrations.source.postgres.legacy.testFixtures
+package io.airbyte.integrations.source.postgres
 
 import io.airbyte.cdk.test.fixtures.legacy.ContainerFactory
 import java.io.IOException
